pyproject.toml
[project]
name = "gh-space-shooter"
version = "0.1.0"
description = "A CLI tool that visualizes GitHub contribution graphs as gamified GIFs"
readme = "README.md"
authors = [
    { name = "zane", email = "czl970721@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
gh-space-shooter = "gh_space_shooter.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.bumpversion]
current_version = "0.0.0"
parse = """(?x)
  (?P<major>0|[1-9]\\d*)\\.
  (?P<minor>0|[1-9]\\d*)\\.
  (?P<patch>0|[1-9]\\d*)
"""
serialize = [
  "{major}.{minor}.{patch}",
]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = false
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = "version = \"{current_version}\""
replace = "version = \"{new_version}\""